requirements.txt
spotipy>=2.22.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
boto3>=1.26.0
requests>=2.28.0
umap-learn>=0.5.3
